projects/ai-news-aggregator/requirements.txt
# AI智能新闻聚合平台 - 简化依赖包
# 核心功能所需的最小依赖

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP客户端
aiohttp==3.9.1
requests==2.31.0

# HTML解析
beautifulsoup4==4.12.2

# RSS处理
feedparser==6.0.10

# 图像处理
Pillow==10.1.0

# API密钥管理
python-dotenv==1.0.0

# 时间处理
python-dateutil==2.8.2

# 类型检查
pydantic==2.5.2